requirements.txt
crewai
textwrap3
langchain-community
duckduckgo_search
streamlit
langchain.tools
amadeus
